requirements.txt
absl-py==0.7.0
astor==0.7.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
numpy==1.16.1
pkg-resources==0.0.0
protobuf==3.6.1
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.31.1
Werkzeug==0.14.1
absl-py==0.7.1
astor==0.7.1
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==3.0.5
numpy==1.16.3
pkg-resources==0.0.0
protobuf==3.7.1
PyYAML==5.1
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tqdm==4.32.1
Werkzeug==0.15.4
